null checks

+++ Assets/Scripts/Storage/EditorStorageManager.cs
@@ -2,7 +2,7 @@ using System.IO;
 using Newtonsoft.Json;
 using UnityEngine;
 
-namespace PlayPerfect.SaveSystem
+namespace PlayPerfect.StorageSystem
 {
     public class EditorStorageManager<T> : IStorageManager<T>
     {
@@ -17,6 +17,18 @@ namespace PlayPerfect.SaveSystem
 
         public void Save(string key, T data)
         {
+            if (string.IsNullOrEmpty(key))
+            {
+                Debug.LogError($"Missing key.");
+                return;
+            }
+            
+            if (data == null)
+            {
+                Debug.LogError($"{nameof(data)} is null.");
+                return;
+            }
+            
             var path = Path.Combine(_saveDirectory, $"{key}.txt");
             var json = JsonConvert.SerializeObject(data);
             File.WriteAllText(path, json);
@@ -24,20 +36,42 @@ namespace PlayPerfect.SaveSystem
 
         public T Load(string key)
         {
+            if (string.IsNullOrEmpty(key))
+            {
+                Debug.LogError($"Missing key.");
+                return default;
+            }
+            
             var path = Path.Combine(_saveDirectory, $"{key}.txt");
-            if (!File.Exists(path)) return default;
+            if (!File.Exists(path))
+            {
+                Debug.LogError($"Path does not exist -> {path}");
+                return default;
+            }
             var json = File.ReadAllText(path);
             return JsonConvert.DeserializeObject<T>(json);
         }
 
         public bool HasKey(string key)
         {
+            if (string.IsNullOrEmpty(key))
+            {
+                Debug.LogError($"Missing key.");
+                return default;
+            }
+
             var path = Path.Combine(_saveDirectory, $"{key}.txt");
             return File.Exists(path);
         }
 
         public void Delete(string key)
         {
+            if (string.IsNullOrEmpty(key))
+            {
+                Debug.LogError($"Missing key.");
+                return;
+            }
+
             var path = Path.Combine(_saveDirectory, $"{key}.txt");
             if (File.Exists(path)) 
                 File.Delete(path);
