+++ Assets/Scripts/Storage/PlayerPrefsStorageManager.cs
@@ -1,12 +1,24 @@
 using Newtonsoft.Json;
 using UnityEngine;
 
-namespace PlayPerfect.SaveSystem
+namespace PlayPerfect.StorageSystem
 {
     public class PlayerPrefsStorageManager<T> : IStorageManager<T>
     {
         public void Save(string key, T data)
         {
+            if (string.IsNullOrEmpty(key))
+            {
+                Debug.LogError($"Missing key.");
+                return;
+            }
+                
+            if (data == null)
+            {
+                Debug.LogError($"{nameof(data)} is null.");
+                return;
+            }
+            
             var json = JsonConvert.SerializeObject(data);
             PlayerPrefs.SetString(key, json);
             PlayerPrefs.Save();
@@ -14,15 +26,33 @@ namespace PlayPerfect.SaveSystem
 
         public T Load(string key)
         {
+            if (string.IsNullOrEmpty(key))
+            {
+                Debug.LogError($"Missing key.");
+                return default;
+            }
+            
             if (!PlayerPrefs.HasKey(key)) return default;
             var json = PlayerPrefs.GetString(key);
             return JsonConvert.DeserializeObject<T>(json);
         }
 
-        public bool HasKey(string key) => PlayerPrefs.HasKey(key);
+        public bool HasKey(string key)
+        {
+            if (!string.IsNullOrEmpty(key))
+                return PlayerPrefs.HasKey(key);
+            Debug.LogError($"Missing key.");
+            return false;
+        }
 
         public void Delete(string key)
         {
+            if (string.IsNullOrEmpty(key))
+            {
+                Debug.LogError($"Missing key.");
+                return;
+            }
+            
             PlayerPrefs.DeleteKey(key);
         }
     }
